first

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -10,7 +10,7 @@
 */
 size_t list_len(const list_t *h)
 {
-unsigned int size = 0;
+size_t size = 0;
 while (h != NULL)
 {
 size++;
